continue splitting modules
