+++ teal/errors.tl
@@ -23,6 +23,34 @@ local record errors
    end
 end
 
+local type Error = errors.Error
+
+function errors.clear_redundant_errors(errs: {Error})
+   local redundant: {integer} = {}
+   local lastx, lasty = 0, 0
+   for i, err in ipairs(errs) do
+      err.i = i
+   end
+   table.sort(errs, function(a: Error, b: Error): boolean
+      local af = assert(a.filename)
+      local bf = assert(b.filename)
+      return af < bf
+             or (af == bf   and (a.y < b.y
+             or (a.y == b.y and (a.x < b.x
+             or (a.x == b.x and (a.i < b.i))))))
+   end)
+   for i, err in ipairs(errs) do
+      err.i = nil
+      if err.x == lastx and err.y == lasty then
+         table.insert(redundant, i)
+      end
+      lastx, lasty = err.x, err.y
+   end
+   for i = #redundant, 1, -1 do
+      table.remove(errs, redundant[i])
+   end
+end
+
 local wk <total>: {errors.WarningKind:boolean} = {
    ["unknown"] = true,
    ["unused"] = true,
